requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorclass==2.2.2
coverage==7.5.3
dnspython==2.6.1
docopt==0.6.2
email_validator==2.1.2
fastapi==0.111.0
fastapi-cache2==0.2.1
fastapi-cli==0.0.4
flake8==7.1.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
ijson==3.2.3
iniconfig==2.0.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
multidict==6.0.5
orjson==3.10.5
packaging==24.1
pendulum==3.0.0
pluggy==1.5.0
pur==7.3.1
pycodestyle==2.12.0
pydantic==2.7.4
pydantic_core
pydeps==1.12.19
pyflakes==3.2.0
Pygments==2.18.0
pytest==8.2.2
pytest-cov==5.0.0
pytest-sugar==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.31.0
responses==0.21.0
rfc3986==1.5.0
rich==13.7.1
setuptools==69.5.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette
stdlib-list==0.10.0
termcolor==2.4.0
terminaltables==3.1.10
time-machine==2.14.1
tree-sitter==0.20.4
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
yarl==1.9.4
